modified
Modified send_queue and receive_queue pointers

+++ mq_lib.h
@@ -5,6 +5,7 @@
 #include <stdlib.h>
 #include <sys/ipc.h>
 #include <sys/msg.h>
+#include <unistd.h>
 
 typedef struct {
     long mtype;
@@ -12,7 +13,7 @@ typedef struct {
 }buffer;
 
 int open_queue (char[], int, int);
-void * receive_queue (int mid, void * msg, size_t som, int attr);
+void * receive_queue (int mid, void * msg, size_t som, long attr);
 void send_queue (int mid, void * msg, size_t som);
 
 #endif
